refactor: Legacy MVC 제거
