+++ app/src/main/java/camp/nextstep/controller/ForwardController.java
@@ -1,0 +1,17 @@
+package camp.nextstep.controller;
+
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Controller
+public class ForwardController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView view(final HttpServletRequest request, final HttpServletResponse response) {
+        return ModelAndView.withJspViewFrom("/index.jsp");
+    }
+}
